requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
cryptography>=41.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0
mcp>=0.1.0
PyJWT>=2.8.0
